# Improve the `logpdf` of `NegativeBinomial`