--- a/src/univariate/discrete/negativebinomial.jl
+++ b/src/univariate/discrete/negativebinomial.jl
@@ -138,35 +138,39 @@ cf(d::NegativeBinomial, t::Real) = laplace_transform(d, -t*im)
 
 # ChainRules definitions
 
+## Callable struct to fix type inference issues caused by captured values
+struct LogPDFNegativeBinomialPullback{D,T<:Real}
+    ∂r::T
+    ∂p::T
+end
+
+function (f::LogPDFNegativeBinomialPullback{D})(Δ) where {D}
+    Δr = Δ * f.∂r
+    Δp = Δ * f.∂p
+    Δd = ChainRulesCore.Tangent{D}(; r=Δr, p=Δp)
+    return ChainRulesCore.NoTangent(), Δd, ChainRulesCore.NoTangent()
+end
+
 function ChainRulesCore.rrule(::typeof(logpdf), d::NegativeBinomial, k::Real)
     # Compute log probability
     r, p = params(d)
-    edgecase = isone(p) && iszero(k)
-    insupp = insupport(d, k)
-    
-    # Primal computation
-    Ω = r * log(p) + k * log1p(-p)
-    if edgecase
-        Ω = zero(Ω)
-    elseif !insupp
-        Ω = oftype(Ω, -Inf)
+    z = xlogy(r, p) + xlog1py(k, -p)
+    if iszero(k)
+        Ω = z
+        ∂r = oftype(z, log(p))
+        ∂p = oftype(z, r/p)
+    elseif insupport(d, k)
+        Ω = z - log(k + r) - logbeta(r, k + 1)
+        ∂r = oftype(z, log(p) - inv(k + r) - digamma(r) + digamma(r + k + 1))
+        ∂p = oftype(z, r/p - k / (1 - p))
     else
-        Ω = Ω - log(k + r) - logbeta(r, k + 1)
+        Ω = oftype(z, -Inf)
+        ∂r = oftype(z, NaN)
+        ∂p = oftype(z, NaN)
     end
 
     # Define pullback
-    function logpdf_NegativeBinomial_pullback(Δ)
-        Δr = Δ * (log(p) - inv(k + r) - digamma(r) + digamma(r + k + 1))
-        Δp = Δ * (r / p - k / (1 - p))
-        if edgecase
-            Δp = oftype(Δp, Δ * r)
-        elseif !insupp
-            Δr = oftype(Δr, NaN)
-            Δp = oftype(Δp, NaN)
-        end
-        Δd = ChainRulesCore.Tangent{typeof(d)}(; r=Δr, p=Δp)
-        return ChainRulesCore.NoTangent(), Δd, ChainRulesCore.NoTangent()
-    end
+    logpdf_NegativeBinomial_pullback = LogPDFNegativeBinomialPullback{typeof(d),typeof(z)}(∂r, ∂p)
 
     return Ω, logpdf_NegativeBinomial_pullback
 end