--- a/test/univariate/discrete/negativebinomial.jl
+++ b/test/univariate/discrete/negativebinomial.jl
@@ -2,11 +2,14 @@ using Distributions
 using Test, ForwardDiff
 using ChainRulesTestUtils
 using FiniteDifferences
+using StatsFuns
 
 # Currently, most of the tests for NegativeBinomial are in the "ref" folder.
 # Eventually, we might want to consolidate the tests here
 
-mydiffp(r, p, k) = r/p - k/(1 - p)
+mydiffp(r, p, k) = iszero(k) ? r/p : r/p - k/(1 - p)
+mydiffr(r, p, k) = iszero(k) ? log(p) : log(p) - inv(k + r) - digamma(r) + digamma(r + k + 1)
+
 @testset "issue #1603" begin
     d = NegativeBinomial(4, 0.2)
     fdm = central_fdm(5, 1)
@@ -21,19 +24,29 @@ mydiffp(r, p, k) = r/p - k/(1 - p)
     @test fdm2(Base.Fix1(cf, d), 0) ≈ -m2
 end
 
-
 @testset "NegativeBinomial r=$r, p=$p, k=$k" for
     p in exp10.(-10:0) .- eps(), # avoid p==1 since it's not differentiable
         r in exp10.(range(-10, stop=2, length=25)),
             k in (0, 1, 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024)
 
     @test ForwardDiff.derivative(_p -> logpdf(NegativeBinomial(r, _p), k), p) ≈ mydiffp(r, p, k) rtol=1e-12 atol=1e-12
+    @test ForwardDiff.derivative(_r -> logpdf(NegativeBinomial(_r, p), k), r) ≈ mydiffr(r, p, k) rtol=1e-12 atol=1e-12
 end
 
 @testset "Check the corner case p==1" begin
-    @test logpdf(NegativeBinomial(0.5, 1.0), 0) === 0.0
-    @test logpdf(NegativeBinomial(0.5, 1.0), 1) === -Inf
-    @test all(iszero, rand(NegativeBinomial(rand(), 1.0), 10))
+    for r in randexp(10)
+        d = NegativeBinomial(r, 1.0)
+        @test @inferred(logpdf(d, 0)) === 0.0
+        @test @inferred(logpdf(d, -1)) === -Inf
+        @test @inferred(logpdf(d, 1)) === -Inf
+        @test all(iszero, rand(d, 10))
+    end
+end
+
+@testset "Check the corner case k==0" begin
+	for r in randexp(5), p in rand(5)
+        @test @inferred(logpdf(NegativeBinomial(r, p), 0)) === xlogy(r, p)
+    end
 end
 
 @testset "rrule: logpdf of NegativeBinomial" begin
@@ -56,4 +69,12 @@ end
     fdm = backward_fdm(5, 1; max_range = r/10)
     test_rrule(logpdf, dist, 0; fdm=fdm)
     test_rrule(logpdf, dist, 0.0 ⊢ ChainRulesTestUtils.NoTangent(); fdm=fdm)
+end
+
+@testset "issue #1582" begin
+    dp = mydiffp(1.0, 1.0, 0.0)
+    @test ForwardDiff.derivative(p -> logpdf(NegativeBinomial(1.0, p), 0.0), 1.0) == dp == 1.0
+
+    dr = mydiffr(1.0, 1.0, 0.0)
+    @test ForwardDiff.derivative(r -> logpdf(NegativeBinomial(r, 1.0), 0.0), 1.0) == dr == 0.0
 end