Uses `xlogy` and `xlog1py` now which take care of the special case `p == 1` automatically. Moreover, for improved numerical accuracy for all cases where `k == 0` (and a bit improved efficiency) there's a new branch for `k == 0`.

Will fix #1582 once the rules for `xlogy` and `xlog1py` in DiffRules and LogExpFunctions are fixed (will open PRs).

Edit: Fixes #1582.