@simsurace 2022-07-05
> Will fix https://github.com/JuliaStats/Distributions.jl/issues/1582

Wouldn't it make sense to add a test for this now?