@simsurace 2022-08-24
Now that https://github.com/JuliaDiff/DiffRules.jl/pull/85 and https://github.com/JuliaStats/LogExpFunctions.jl/pull/57 have been merged, this could be revived, right?