@devmotion 2022-08-31
FYI I applied similar changes also to the reverse-mode rule, and moved computations of constants outside of the pullback. That leads to a significant performance improvement of the pullback function (of course, only matters if it is evaluated multiple times):

## On master

```julia
julia> using Distributions, ChainRulesCore, BenchmarkTools

julia> d = NegativeBinomial(0.8, 0.5);

julia> x = 1;

julia> @btime rrule($logpdf, $d, $x);
  45.421 ns (0 allocations: 0 bytes)

julia> _, pb = rrule(logpdf, d, x);

julia> @btime $pb($(randn()));
  36.472 ns (0 allocations: 0 bytes)
```

## With this PR

```julia
julia> using Distributions, ChainRulesCore, BenchmarkTools

julia> d = NegativeBinomial(0.8, 0.5);

julia> x = 1;

julia> @btime rrule($logpdf, $d, $x);
  77.026 ns (0 allocations: 0 bytes)

julia> _, pb = rrule(logpdf, d, x);

julia> @btime $pb($(randn()));
  1.501 ns (0 allocations: 0 bytes)
```